requirements.txt
Flask>=3.0.0
python-dotenv>=1.0.0
langchain-google-genai>=2.1.5
browser-use>=0.2.6  # Actualizado a la versión más reciente
langchain-anthropic>=0.3.15  # Para Claude
playwright>=1.52.0  # Para automatización del navegador
langchain-core>=0.3.64  # Dependencia core de LangChain
langchain-openai>=0.3.21  # Para OpenAI
pydantic>=2.11.5  # Para modelos de datos
pyautogui>=0.9.54  # Para capturas de pantalla
requests>=2.32.3  # Para API calls
pillow>=9.0.0  # Para manipulación de imágenes
asyncio>=3.4.3  # Para manejo asíncrono
orjson>=3.0.0  # Para manejo eficiente de JSON
beautifulsoup4>=4.9.0  # Para parseo HTML
anthropic>=0.52.0  # SDK oficial de Anthropic

# Dependencias adicionales requeridas por browser-use 0.2.6
anyio>=4.9.0
faiss-cpu>=1.11.0
langchain>=0.3.25
langchain-aws>=0.2.24
langchain-deepseek>=0.1.3
langchain-ollama>=0.3.3
mem0ai>=0.1.106
patchright>=1.52.5
posthog>=3.7.0
psutil>=7.0.0
pyperclip>=1.9.0
screeninfo>=0.8.1
httpx>=0.28.1
typing-extensions>=4.12.2
uuid7>=0.1.0

# Opcional: Para ejecución de tareas en segundo plano (si se decide usar)
# celery>=5.0
# redis>=4.0
# flower>=1.0 